move contract to gridcoin folder and change includes to suit
