+++ src/gridcoin/contract.h
