+++ src/gridcoin/polls.cpp
@@ -6,10 +6,10 @@
 #include "contract.h"
 #include "rpcclient.h"
 #include "rpcserver.h"
-#include "gridcoin/appcache.h"
-#include "gridcoin/cpid.h"
+#include "appcache.h"
+#include "cpid.h"
 #include "init.h" // for pwalletMain
-#include "gridcoin/block.h"
+#include "block.h"
 
 double GetTotalBalance();
 std::string TimestampToHRDate(double dtm);
