+++ src/gridcoin/polls.h
