+++ src/main.cpp
@@ -26,7 +26,7 @@
 #include "gridcoin/backup.h"
 #include "gridcoin/appcache.h"
 #include "gridcoin/tally.h"
-#include "gridcoin/contract/contract.h"
+#include "gridcoin/contract.h"
 
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/fstream.hpp>
